requirements.txt
fastapi
uvicorn[standard]
python-multipart
fastapi-login
passlib[bcrypt]
python-jose
jinja2
rich
config
mariadb
sqlalchemy
alembic
aiofiles
babel
Flask-Script
pydantic[email]
python-dateutil
fastapi-limiter
phpserialize
sqlalchemy_utils
python-dotenv